Moved repositories service to core to prevent module dependencies.
(cherry picked from commit 0144123)
