+++ src/js/angular/export/app.js
@@ -2,7 +2,7 @@ import 'angular/core/services';
 import 'angular/core/directives';
 import 'angular/export/controllers';
 import 'angular/directives/paginations';
-import 'angular/repositories/services';
+import 'angular/core/services/repositories.service';
 import 'lib/FileSaver-patch';
 
 const modules = [
@@ -12,7 +12,7 @@ const modules = [
     'ngCookies',
     'graphdb.framework.impex.export.controllers',
     'graphdb.framework.directives.paginations',
-    'graphdb.framework.repositories.services',
+    'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives'
 ];
