+++ src/js/angular/import/app.js
@@ -2,7 +2,7 @@ import 'angular/core/services';
 import 'angular/core/directives';
 import 'angular/import/controllers';
 import 'angular/import/directives';
-import 'angular/repositories/services';
+import 'angular/core/services/repositories.service';
 import 'ng-file-upload/dist/ng-file-upload.min';
 import 'ng-file-upload/dist/ng-file-upload-shim.min';
 
@@ -13,7 +13,7 @@ const modules = [
     'ui.bootstrap',
     'graphdb.framework.impex.import.directives',
     'graphdb.framework.impex.import.controllers',
-    'graphdb.framework.repositories.services',
+    'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives'
 ];
