+++ src/js/angular/repositories/app.js
@@ -2,7 +2,7 @@ import 'angular/core/services';
 import 'angular/core/directives';
 import 'angular/core/controllers';
 import 'angular/repositories/controllers';
-import 'angular/repositories/services';
+import 'angular/core/services/repositories.service';
 import 'angular/settings/app';
 import 'ng-file-upload/dist/ng-file-upload.min';
 import 'ng-file-upload/dist/ng-file-upload-shim.min';
@@ -13,7 +13,7 @@ const modules = [
     'ngCookies',
     'ngRoute',
     'graphdb.framework.repositories.controllers',
-    'graphdb.framework.repositories.services',
+    'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives',
     'graphdb.framework.core.controllers',
     'graphdb.framework.settings'
