+++ src/js/angular/repositories/controllers.js
@@ -1,5 +1,4 @@
 import 'angular/core/services';
-import 'angular/repositories/services';
 
 const filenamePattern = new RegExp('^[a-zA-Z0-9-_]+$');
 const numberPattern = new RegExp('[0-9]');
@@ -22,7 +21,7 @@ const staticRulesets = [
 const modules = [
     'ngCookies',
     'ui.bootstrap',
-    'graphdb.framework.repositories.services',
+    'graphdb.framework.core.services.repositories',
     'toastr',
     'ngFileUpload'
 ];
