+++ src/js/angular/resources/app.js
@@ -1,7 +1,7 @@
 import 'angular/core/services';
 import 'angular/core/directives';
 import 'angular/resources/controllers';
-import 'angular/repositories/services';
+import 'angular/core/services/repositories.service';
 import 'd3/build/d3';
 import 'lib/nvd3/angular-nvd3';
 
@@ -10,7 +10,7 @@ const modules = [
     'ui.bootstrap',
     'nvd3',
     'graphdb.framework.jmx.resources.controllers',
-    'graphdb.framework.repositories.services',
+    'graphdb.framework.core.services.repositories',
     'graphdb.framework.core.directives'
 ];
